% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-command.R
\name{machine_command}
\alias{machine_command}
\title{Run arbitrary docker-machine commands.}
\usage{
machine_command(cmd, ..., machine = NULL, echo = TRUE)
}
\arguments{
\item{cmd}{[\code{character(1)}]\cr A docker-machine CLI command to run}

\item{...}{\code{[name=value pairs}]\cr Name value pairs of extra options
to pass to \code{cmd}. All option names will be prefixed with \code{--} before being
passed to docker-machine.}

\item{machine}{\code{[character: NULL}]\cr The name of the docker-machine
to run the command on, if applicable.}
}
\description{
Run arbitrary docker-machine commands.
}
