% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-config.R
\name{machine_config}
\alias{machine_config}
\title{Print the connection configuration for a machine}
\usage{
machine_config(machine, ...)
}
\arguments{
\item{machine}{[\code{character}, \code{machine}]\cr A \code{machine}, or \code{character} vector
of machine names.}

\item{...}{\code{[name=value pairs}]\cr Name value pairs of extra options
to pass to \code{cmd}. All option names will be prefixed with \code{--} before being
passed to docker-machine.}
}
\description{
Print the connection configuration for a machine
}
