% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-ip.R
\name{machine_ip}
\alias{machine_ip}
\title{Query the IP of a machine instance or cluster}
\usage{
machine_ip(machine)
}
\arguments{
\item{machine}{[\code{character}, \code{machine}]\cr A \code{machine}, or \code{character} vector
of machine names.}
}
\description{
Query the IP of a machine instance or cluster
}
