% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-kill.R
\name{machine_kill}
\alias{machine_kill}
\title{Kill a machine}
\usage{
machine_kill(machine)
}
\arguments{
\item{machine}{[\code{character}, \code{machine}]\cr A \code{machine}, or \code{character} vector
of machine names.}
}
\description{
Killing a machine abruptly stops it. Using \code{\link[=machine_stop]{machine_stop()}} and
\code{\link[=machine_rm]{machine_rm()}} is advised instead if possible.
}
