% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-rm.R
\name{machine_rm}
\alias{machine_rm}
\title{Remove a machine}
\usage{
machine_rm(machine, force = FALSE)
}
\arguments{
\item{machine}{[\code{character}, \code{machine}]\cr A \code{machine}, or \code{character} vector
of machine names.}

\item{force}{[\code{logical(1)}]\cr Remove local configuration even if
cloud/virtual machine cannot be removed.}
}
\description{
This removes the local reference and deletes it on the cloud provider or
virtualization management platform.
}
