% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-stop.R
\name{machine_stop}
\alias{machine_stop}
\title{Stop a machine}
\usage{
machine_stop(machine)
}
\arguments{
\item{machine}{[\code{character}, \code{machine}]\cr A \code{machine}, or \code{character} vector
of machine names.}
}
\description{
Gracefully stop a machine.
}
